Мелкие правки по дизайну, виджет, help
Дописал отступы в диалогах перед EditText-ами
Отладил минимальные размеры кнопок
Прописал стиль для AutocompleteTextView
Добавил help, выплывающий при первом запуске
Добавил виджеты
Добавил md-файл и скриншоты

+++ app/src/main/java/com/ultra/shopperlights2/Activities/EditHistoryActivity.java
@@ -101,7 +101,7 @@ public class EditHistoryActivity extends AppCompatActivity
 		{
 		for(Purchase purchase : p)
 			{
-			if(purchase.getPrice()==0 && purchase.isCompleted() )
+			if(purchase.getPrice()==0 && purchase.isCompleted() && purchase.getProducts().size()==0)
 				App.session.getPurchaseDao().delete(purchase);
 			}
 		}
